Crear microservicios y comunicación asíncrona, CRUD para Movimientos y Cuentas

+++ ms-clientes-personas/src/main/java/com/rommelchocho/ms_clientes_personas/controller/ClienteController.java
@@ -3,11 +3,14 @@ package com.rommelchocho.ms_clientes_personas.controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.rommelchocho.ms_clientes_personas.dto.ClienteDto;
 import com.rommelchocho.ms_clientes_personas.entity.Cliente;
 import com.rommelchocho.ms_clientes_personas.service.ClienteService;
 
 import java.util.List;
+import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,36 +31,44 @@ public class ClienteController {
 
     // Crear un nuevo Cliente
     @PostMapping
-    public ResponseEntity<Cliente> createCliente(@RequestBody Cliente cliente) {
-        Cliente nuevoCliente = clienteService.createCliente(cliente);
-        return ResponseEntity.ok(nuevoCliente);
-
+    public ResponseEntity<?> createCliente(@RequestBody Cliente cliente) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(clienteService.createCliente(cliente));
     }
 
     // Obtener todos los Clientes
     @GetMapping
-    public ResponseEntity<List<Cliente>> getAllClientes() {
+    public ResponseEntity<List<ClienteDto>> getAllClientes() {
         return ResponseEntity.ok(clienteService.getAllClientes());
-
     }
 
     // Obtener un Cliente por ID
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> getClienteById(@PathVariable Long id) {
-        return ResponseEntity.ok(clienteService.getClienteById(id));
+    public ResponseEntity<?> getClienteById(@PathVariable Long id) {
+        Optional<ClienteDto> cliente = clienteService.getClienteById(id);
+        if (cliente.isPresent()) {
+            return ResponseEntity.ok(cliente.orElseThrow());
+        }
+        return ResponseEntity.notFound().build();
     }
 
     // Actualizar un Cliente
     @PutMapping("/{id}")
-    public ResponseEntity<Cliente> updateCliente(@PathVariable Long id, @RequestBody Cliente cliente) {
-        Cliente actualizado = clienteService.updateCliente(id, cliente);
-        return ResponseEntity.ok(actualizado);
+    public ResponseEntity<?> updateCliente(@PathVariable Long id, @RequestBody Cliente cliente) {
+        Optional<ClienteDto> clienteUpdated = clienteService.updateCliente(id, cliente);
+        if (clienteUpdated.isPresent()) {
+            return ResponseEntity.status(HttpStatus.CREATED).body(clienteUpdated.orElseThrow());
+        }
+        return ResponseEntity.notFound().build();
     }
 
     // Eliminar un Cliente
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
-        clienteService.deleteCliente(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<?> deleteCliente(@PathVariable Long id) {
+        Optional<ClienteDto> cliente = clienteService.getClienteById(id);
+        if (cliente.isPresent()) {
+            clienteService.deleteCliente(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
